+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * github-175: '--jobs' parameter starts way to many parallel jobs
 * github-176: start-/end-date in report should not be the same
+* github-177: Implement generic SQL schema check
 
 
 # Version 0.24.0 - 2022-04-05